Pred网络输出acc/delta，用自行车模型计算vtan(delta)并积分生成轨迹，在loss中与真实轨迹比较，但是可能因为v/delta梯度互相影响，且包含双重积分，迭代基本完全发散

+++ data.py
@@ -29,7 +29,7 @@ class InteDataset(Dataset):
         data = pickle.load(f, encoding="latin1")
         data['feats'] = data['feats'].astype('float32')
         data['orig'] = data['orig'].astype('float32')
-        data['theta'] = data['theta'].astype('float32')
+        # data['theta'] = data['theta'].astype('float32')
         data['trajs'] = data['trajs'].astype('float32')
 
         graph = preprocess(to_long(gpu(data['graph'])), 6) #self.config['cross_dist']为6
